Passport

+++ app/Http/Controllers/Api/SubjectController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Subject;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Validation\ValidationException;
 
 class SubjectController extends Controller
@@ -17,7 +19,7 @@ class SubjectController extends Controller
      */
     public function index()
     {
-        return response()->json(Subject::all(), 200);
+        return response()->json(Subject::all(), Response::HTTP_OK);
     }
 
     /**
@@ -33,7 +35,7 @@ class SubjectController extends Controller
 
         $subject = Subject::create($data);
 
-        return response()->json($subject, 201);
+        return response()->json($subject, Response::HTTP_CREATED);
     }
 
     /**
@@ -44,7 +46,7 @@ class SubjectController extends Controller
      */
     public function show(Subject $subject)
     {
-        return response()->json($subject, 200);
+        return response()->json($subject, Response::HTTP_OK);
     }
 
     /**
@@ -61,7 +63,7 @@ class SubjectController extends Controller
 
         $subject->update($data);
 
-        return response()->json($subject, 200);
+        return response()->json($subject, Response::HTTP_OK);
     }
 
     /**
@@ -69,12 +71,18 @@ class SubjectController extends Controller
      *
      * @param Subject $subject
      * @return JsonResponse
-     * @throws \Exception
+     * @throws Exception
      */
     public function destroy(Subject $subject)
     {
+        // if the subject has tasks, update tasks first
+        if ($subject->tasks->all())
+        {
+            return response()->json(null, Response::HTTP_CONFLICT);
+        }
+
         $subject->delete();
 
-        return response()->json(null, 204);
+        return response()->json(null, Response::HTTP_NO_CONTENT);
     }
 }
